DROTH-3121 reorganize vvh and kgv client for future deletion of vvh code
